Experiments are found even with running watcher.
To successfully test this feature, ExperimentsWatcher had to be
reworked, because popen/pclose solution was not able to read
only part of the output and caused infinite loop.
This solutions uses output redirection, so we can read ouput already
written to the file by file_get_contents and then kill the process.